chore: [WLEO-386] Update proximity implementation to version 0.3.0 (#81)
> [!WARNING]
> This is a rough implementation only meant for testing purposes. It
currently works only on Android. If somethings wrong happens during the
flow, close the page with the QR code and open it again as the reload
button doesn't seem to generate a new QR code
This PR updates `io-react-native-proximity` and the proximity
implementation aligning it with the latest changes.
Include a summary of the changes.

- Update `io-react-native-proximity` to version `0.3.0`;
- Update the proximity implementation in the `useProximity` hook which
includes changes on how we generate the accepted fields JSON object
which is currently hardcoded and some minor modifications on how we
provide documents to the native methods via an array.
`yarn
cd ios && bundler exec pod install && cd ..
yarn react-native run-android`
Test a presentation flow after obtaining an MDL with the reference
implementation verifier app.
Describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce.