chore: [WLEO-397] Add URL wildcard intent Android (#82)
## Short description
This PR adds the intent to open `https` urls with the `hosts` wildcard
in order to open the provided URL, if any, after a presentation flow.
## List of changes proposed in this pull request

- Update the `AndroidManifest` with the required intent;
- Update the toast type which was `success` instead of `error` if the
URL can't be opened;
- Update the generic message locale string.

## How to test
Test a presentation flow which provides a `redirect_uri` at the end of
the flow. The browser should be opened.