refactor: [WLEO-419] Proximity implementation (#83)
> [!WARNING]
> This breaks the remote presentation.

## Short description
This PR refactors the proximity implementation, which is still a very
basic one, according to `io-react-native-proximity` version `0.5.0`.
Due to the requirement of the newest version of `io-react-native-cbor`,
the `io-react-native-wallet` had to be bumped as well, causing the
remote presentation flow to break. It will be fixed in a different pull
request.

## List of changes proposed in this pull request
- Update `io-react-native-proximity`, `io-react-native-wallet` and
`io-react-native-cbor`;
- Refactor the proximity implementation according to the new
`io-react-native-proximity` interface along with some fixes;
- Update the requested permission based on the SDK version on Android
and remove the location permission for iOS;
- Simplifies the logbox management with a custom hook.

## How to test
Test a proximity flow on both Android (reference implementation app as
verifier) and iOS (Scytales mValid app as verifier).