fix: [WLEO-460] Add WIA expiration check (#88)
> [!CAUTION]
> This needs an app reinstallation in order to work as we don't
currently manage store migrations.

## Short description

Include a summary of the changes.

## List of changes proposed in this pull request

- Update the `getAttestation` saga which now checks if an existing
wallet instance attestation exists and if it does, it checks its
validity before trying to obtain a new one. It returns the existing one
or the newly generated one which is also set in the attestation state;
- Remove the wallet instance attestation request when creating a new
wallet instance;
- Add a call to `getAttestation` whenever required, thus when obtaining
the PID or a generic credential;
- Remove the wallet instance attestation existence from the presentation
flow as it's not needed;
- Add the missing persistence for the `instance` slice;
- Refactor the `lifecycle` persistence by moving it directly in its
slice instead persisting the whole wallet slice and using the whitelist
parameter;

## How to test
- Uninstall the app if you have it and reinstall it;
- Try to obtain a PID and a credential.