deps: [WLEO-440] Update React Native to version 0.78 (#85)
## Short description
This PR updates React Native to version 0.78.

## List of changes proposed in this pull request
- Update React Native to version 0.78 according to the [upgrade
helper](https://react-native-community.github.io/upgrade-helper/?from=0.75.4&to=0.78.2&package=IoEudiwApp&name=IoEudiwApp);
- Update dependencies to support the new RN verision;
- Update the ref type for some components as required by the new type.

## How to test
Run `yarn nuke` and then follow the instructions in the repository
[README.md](https://github.com/pagopa/io-eudiw-app/blob/main/README.md).

Run a build on both iOS and Android and check for regressions by
obtaining a credential, presenting it remotely and in proximity.