feat: [WLEO-425,WLEO-448] Implemented Proximity flow (#87)
## Short description

This PR adds the necessary UI to handle a proximity presentation.

## List of changes proposed in this pull request

- Created `proximity` saga.
- Added _Show QR_ button on the _Credential Previews_ of _mDoc_
credentials that triggers a modal with the Proximity QR Code.
- Added `PresentationProximityPreview` screen that displays the claims
that are going to be shared.
- Linked _Show QR_ screen to the new proximity flows and logic.

## How to test

0. Start from a clean install of the application (Design system has been
updated, so a `yarn install` invocation is required).
1. Obtain an _mDL_ and an _HealthID_.
2. Try some proximity presentations, covering as many cases as wanted.
Example of such cases are:
a. From the _Show QR_ screen in the Wallet Home, start a presentation
and share all data, the presentation **should** end successfully.
b. From the _Show QR_ screen in the Wallet Home, start a presentation
and share partial data, the presentation **should** end successfully.
c. From the _Show QR_ screen in the Wallet Home, start a presentation
and cancel it, the presentation **should not** end successfully.
d. From the _Show QR_ screen in the Wallet Home, start a presentation
and refuse to authenticate, the presentation **should not** end
successfully.
e. From the modal inside the _Credential Details_ screen of an _mDoc_
credential, start a presentation and share all data, the presentation
**should** end successfully.
f. From the modal inside the _Credential Details_ screen of an _mDoc_
credential, start a presentation and share partial data, the
presentation **should** end successfully.
g. From the modal inside the _Credential Details_ screen of an _mDoc_
credential, start a presentation and cancel it, the presentation
**should not** end successfully.
h. From the modal inside the _Credential Details_ screen of an _mDoc_
credential, start a presentation and refuse to authenticate, the
presentation **should not** end successfully.
i. Open the _Show QR_ screen, close it, and do a presentation from the
_Credential Preview_ screen and viceversa, the presentation **should**
end successfully.

The tests should be run on iOS and Android.