feat: [WLEO-450] Add certificates for proximity verifiers (#91)
## Short description
This PR adds the proximity verifier certificates lists. 
Include a summary of the changes.

## List of changes proposed in this pull request

- Add an array with the entity name and its certificate;
- Sends the certificates as parameter in the `Proximity.start` method;
- Add the `isAuthenticated` flag in the store along with the descriptor
which is set to true only if all the `isAuthenticated` flags are set to
true for each requested credential, false otherwise;
- Add a component which shows an alert which indicates wether or not the
verifier app is authenticated;
- Update locales.

## How to test
This can't be tested at the moment as we don't have any verifier
available for these certificates.