fix: [WLEO-495] Inconsistency state on iOS after app removal (#95)
## Short description
This PR fixes an inconsistency state on iOS after app removal. This is
caused by the keychain surviving an app removal.

## List of changes proposed in this pull request

- Add handling of `preferencesReset` and `resetLifecycle` actions where
missing. This ensure the state is properly cleared when these two
actions are dispatched;
- Dispatch a `resetLifecycle()` action after the onboarding is
completed. This ensures the wallet state in cleaned properly, including
any value persisted in the keychain;
- Remove the `resetCredentials()` action after creating a new wallet
instance as the cleanup already happened;
- Change the logic associated with the remove credential button in the
credential detail. It now dispatches `resetLifecycle` if the credential
is a PID which properly cleans the whole wallet state.

## How to test
This can be tested in these scenarios: 
- Activate a wallet from scratch and when prompted to enter the
credentials to obtain a PID, cancel the flow. Remove the app and install
it again, then try to activate the wallet as usual and obtain a
credential;
- From an active wallet, remove the app and install it again. Try to
obtain a wallet again and obtain a credential;
- Try to remove the PID from the wallet by navigating in its details
page and try to activate the it again;
- With an active wallet, open the settings menu and reset the wallet.