feat: [WLEO-499] Added a new component to render string array claims (#97)
## Short description

This PR adds a `StringArrayItem` component that handles the rendering of
a claim which consists of an array of string values in the context of
the `CredentialClaims` component.

## List of changes proposed in this pull request

- `ts/features/wallet/components/credential/CredentialClaims.tsx` :
Added the aforementioned new `StringArrayItem` component, which renders,
in preview, at most the first two elements contained in the claim, and,
in case there are more, displays a modal to show the list of all the
claim's values.

- `ts/hooks/useBottomSheet.tsx` : added a new
`maxDynamicContentSizePercent` prop which is a percentage applied to the
value used in the `maxDynamicContentSize` prop of the `BottomSheetModal`
component.

## How to test

Obtain a PID and an FBK badge and assert the following:

> **_Note_**: the checks on the credential claims fields are based on
the fact that, for the time being, these claims are string array claims
that contain a single element, two elements or more than two elements,
but in time their values might change.

- The `nationality` field should be displayed correctly in both the
PID's `PresentationCredentialDetails` and the `CredentialTrust` screens.
More specifically, the *Show* button on the right **should not** be
present.
- In the FBK Badge's `PresentationCredentialDetails` screen the
`Locations` and `Qualifications` entries **should** display a
comma-separated list of two values and an ellipsis, a *Show* button
**should** be present on the right and, when clicked, a
`BottomSheetModal` containing all the values in the claim **should** be
shown. On top of that, the `Benefits` screen **should** display a
comma-separated list of two items and there **should not** be an *Show*
button on the right.
- (testable only by modifying the source code or changing the credential
on the QEAA Provider side) add a number of claims that would cause an
overflow on any of the aforementioned claims and **assert** that the
`BottomSheetModal` does not grow for more than approximatively the 75%
of the screen height, and **assert** that all elements are visible by
scrolling. (**_Note_**: this has been tested on local builds).