fix: [WLEO-560] Separated credential type from corresponding EC configuration ID (#98)
## Short description

This PR fixes the credential issuance and presentation mechanisms to
allow separation of the credentialType (SDJWT `vct` or MDoc `scope`)
from the identifier used in the EC configuration to distinguish the
available credentials.

With this differentiation, the credential configuration ID is used only
in the credential issuance sagas, while the credential type will be used
for all localization, rendering and credential identification, replacing
in fact the credential configuration ID.

## List of changes proposed in this pull request

- `ts/features/wallet/utils/credentials.ts` : Changed
`wellKnownCredential`'s values to credential types instead of Entity
Configuration IDs and added `wellKnownCredentialConfigurationIDs` to
replace its use.
- `ts/features/wallet/saga/(pid|credential).ts`: Made the flow start
with the EC credential configuration ID, added extraction of credential
type from EC.
- `ts/features/wallet/store/credentialIssuance.ts` : Added
`requestedCredentialType` property to use the EC extracted credential
type for the remainder of the issuance session.
- Adapted components to the changes above.

## How to test

Try to run a proximity presentation with an FBK badge capable verifier,
and check that:
1. the information is displayed correctly,
2. the verifier receives the requested credential.