feat: [WLEO-687] Migration to `io-app-design-system:5.11.13-0` (#106)
## Short description

This PR handles the migration to the latest `io-app-design-system`
version.

## List of changes proposed in this pull request

- Migrated io-app-design system to latest version, 
- Reintroduced deprecated `RawAccordion` component under the new name of
`ClaimsSelector` due to incompatibility with `eudiw`'s needs,
- Created a new `preview` Android build profile to create an unsigned
release apk for testing.

## How to test

Install the app and compare the various screens with the current `eudiw`
version. The screens should be the same, except for the following :

- For the `ClaimsSelector` component, now on landing on the presentation
screens the accordion is already open.
- The `ForceScrolldownView` scroll to bottom button should work better,
especially in the Credential Preview Screen.

In order to test both apps simultaneously, a `preview` Android build
config has been added, which bundles an apk that doesn't depend on the
debug metro server and doesn't require signing.

In case the preview config is deemed to be useless, it will be removed.